añade comprobaciones en los body de las peticiones

+++ src/controllers/cliente.js
@@ -2,6 +2,9 @@ const Cliente = require("../models/cliente")
 
 // crear cliente (C)
 async function crearCliente(req,res,next){
+    if(!req.body.dni || !req.body.nombre){
+        return res.status(400).json({msg: "Faltan datos requeridos"})
+    }
     const nuevo = new Cliente(req.body)
     await nuevo.save()
     res.json(nuevo)
@@ -15,12 +18,22 @@ async function obtenerClientes(req,res,next){
 
 // actualizar cliente (U)
 async function actualizarCliente(req,res,next){
+    if(!req.params.id){
+        return res.status(400).json({msg: "Parámetro id requerido"})
+    }
+
+    if(!req.body.direccion){
+        return res.status(400).json({msg: "Faltan datos requeridos: direccion"})
+    }
     await Cliente.findByIdAndUpdate(req.params.id, req.body)
     res.json({msg: "Cliente actualizado"})
 }
 
 // borrar cliente (D)
 async function eliminarCliente(req, res,next){
+    if(!req.params.id){
+        return res.status(400).json({msg: "Parámetro id requerido"})
+    }
     await Cliente.findByIdAndDelete(req.params.id)
     res.json({msg: "Cliente eliminado"})
 }
