+++ src/controllers/venta.js
@@ -2,6 +2,9 @@ const Venta = require("../models/venta")
 
 // crear venta (C)
 async function crearVenta(req,res,next){
+    if(!req.body.vendedor || !req.body.comprador || !req.body.vehiculo || !req.body.importe){
+        return res.status(400).json({msg: "Faltan datos requeridos"})
+    }
     const nuevo = new Venta(req.body)
     await nuevo.save()
     res.json(nuevo)
@@ -19,12 +22,23 @@ async function obtenerVentas(req,res,next){
 
 // actualizar venta (U)
 async function actualizarVenta(req,res,next){
+    if(!req.params.id){
+        return res.status(400).json({msg: "Parámetro id requerido"})
+    }
+
+    if(!req.body.importe){
+        return res.status(400).json({msg: "Faltan datos requeridos: importe"})
+    }
+
     await Venta.findByIdAndUpdate(req.params.id, req.body)
     res.json({msg: "Venta actualizado"})
 }
 
 // borrar venta (D)
 async function eliminarVenta(req, res,next){
+    if(!req.params.id){
+        return res.status(400).json({msg: "Parámetro id requerido"})
+    }
     await Venta.findByIdAndDelete(req.params.id)
     res.json({msg: "Venta eliminado"})
 }
